feat: add command buffers & rendering
Completed all the swapchain and graphics pipeline code to a point where
a triangle can be drawn on the screen. Includes triple-buffering and
synchronization between the CPU and GPU.

One important note is that the shader loading functions had to be
rewritten to load data from a spirv file as rust-gpu was not working for
whatever reason. This is only a temporary change though and rust-gpu
will be implemented eventually.