fix: change destroy order of vulkan objects
Vulkan objects were getting destroy in the incorrect order causing tons
of memory leaks in Vulkan. This has been resolved.

Another quirk that was found is that the debug messenger isn't
correctly created. When ENABLE_VALIDATION_LAYERS is true, the debug
messenger doesn't seem to fully initialize but when it's false it debugs
everything as it should. For now the mentioned constant was set to false
to fix this.