feat: add swapchain recreation & window resizing
Resizing the window no longer crashes the app and will instead cause the
swapchain to be recreated to match the new window size. The render pass
is also re-created, even if it's compatible with the new swapchain.

Since the graphics pipeline is re-created every frame the window is
being resized, rust-gpu recompiles the shader crates resulting in
noticeable tearing/lag as you resize a window.