feat: add push constants
Added the ability for shaders to have push constants, for testing it
only includes an offset vec2. However, when trying to use rust-gpu the
push constants would not work but when switching to using GLSL and
compiling it to spirv using glslc the push constants work. Hence the
hacky vertex_shader_module.