fix: cast &str to ptr correctly on Linux
When creating a debug build and casting the VALIDATION_LAYERS to
pointers so that it can be passed through to the instance create info
struct, it would error as the layer could not be found.

Turns out the pointers weren't being gotten correctly and so it was
changed to match the approach taken in the logical device where it is
first cast to a CString so that we can get a native c_char pointer
instead of forcing a conversion to a i8 pointer.