feat: add debug layers to define validation layers
The user, when creating the instance, can pass in an array of debug
layers they want to enable. These debug layers are just an enum wrapper
for Vulkan validation layers.

If no debug layers are passed in then a Debugger won't be created and no
validation layers will be enabled.